Extract useSignalValue and extension hooks into shared module
Create utils/useExtensionHooks.ts with three exports:
- useSignalValue: primitive hook for subscribing to any ReadonlySignal
- useExtensionSignalValue: extension-aware wrapper using useSignalValue
- useExtensionComponent: workaround for upstream Output invariance

useAI now uses useSignalValue directly on the output signals.
ToolbarPlugin imports useSignalValue from the shared module.
Editor.tsx imports useExtensionComponent and useExtensionSignalValue
from the shared module instead of defining them locally.
