+++ examples/agent-example/src/Editor.tsx
@@ -10,34 +10,23 @@ import {TabIndentationExtension} from '@lexical/extension';
 import {HistoryExtension} from '@lexical/history';
 import {ContentEditable} from '@lexical/react/LexicalContentEditable';
 import {LexicalExtensionComposer} from '@lexical/react/LexicalExtensionComposer';
-import {useExtensionDependency} from '@lexical/react/useExtensionComponent';
 import {RichTextExtension} from '@lexical/rich-text';
 import {
   $createParagraphNode,
   $createTextNode,
   $getRoot,
-  type AnyLexicalExtension,
   defineExtension,
-  type LexicalExtensionOutput,
 } from 'lexical';
 
 import {AIExtension} from './ai/AIExtension';
-import {useExtensionSignalValue} from './ai/useAI';
 import {OrgNodeExtension} from './nodes/OrgNode';
 import {PersonNodeExtension} from './nodes/PersonNode';
 import {PlaceNodeExtension} from './nodes/PlaceNode';
 import {ToolbarExtension} from './plugins/ToolbarPlugin';
-
-// TODO: submit fix upstream for useExtensionComponent to handle
-// extensions with register (Output invariance due to ExtensionRegisterState)
-function useExtensionComponent<
-  Extension extends AnyLexicalExtension,
-  Output extends LexicalExtensionOutput<Extension> & {
-    Component: React.ComponentType;
-  },
->(extension: Extension): Output['Component'] {
-  return (useExtensionDependency(extension).output as Output).Component;
-}
+import {
+  useExtensionComponent,
+  useExtensionSignalValue,
+} from './utils/useExtensionHooks';
 
 const theme = {
   heading: {
